Create separate component for message list sort icon
