+++ graylog2-web-interface/src/views/components/widgets/MessageTable.jsx
@@ -2,8 +2,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import * as Immutable from 'immutable';
-import styled, { css, type StyledComponent } from 'styled-components';
-import { isEmpty } from 'lodash';
+import styled from 'styled-components';
 
 import { AdditionalContext } from 'views/logic/ActionContext';
 import MessageFieldsFilter from 'logic/message/MessageFieldsFilter';
@@ -11,9 +10,7 @@ import FieldType from 'views/logic/fieldtypes/FieldType';
 import FieldTypeMapping from 'views/logic/fieldtypes/FieldTypeMapping';
 import MessagesWidgetConfig from 'views/logic/widgets/MessagesWidgetConfig';
 import CustomPropTypes from 'views/components/CustomPropTypes';
-import Direction from 'views/logic/aggregationbuilder/Direction';
-import { Icon } from 'components/common';
-import SortConfig from 'views/logic/aggregationbuilder/SortConfig';
+import MessagesSortIcon from 'views/components/widgets/MessagesSortIcon';
 
 import { RefreshActions } from 'views/stores/RefreshStore';
 
@@ -118,19 +115,6 @@ const Table = styled.table`
   }
 `;
 
-const SortIcon: StyledComponent<{fieldSortActive: boolean}, {}, HTMLButtonElement> = styled.button(({ fieldSortActive }) => {
-  const color = fieldSortActive ? '#333' : '#bdbdbd';
-
-  return css`
-    border: 0;
-    background: transparent;
-    color: ${color};
-
-    padding: 5px;
-
-    cursor: pointer;
-  `;
-});
 
 const TableHead = styled.thead`
   background-color: #eee;
@@ -186,29 +170,6 @@ class MessageTable extends React.Component<Props, State> {
     expandedMessages: Immutable.Set<string>(),
   };
 
-  _changeSort = (config: MessagesWidgetConfig, fieldName: string, fieldSortDirection: ?Direction) => {
-    const { onConfigChange, setLoadingState } = this.props;
-    let newSort;
-
-    setLoadingState(true);
-
-    switch (fieldSortDirection) {
-      case Direction.Ascending:
-        newSort = config.sort.map(sort => sort.toBuilder().direction(Direction.Descending).build());
-        break;
-      case Direction.Descending:
-        newSort = config.sort.map(sort => sort.toBuilder().direction(Direction.Ascending).build());
-        break;
-      default:
-        newSort = [new SortConfig('pivot', fieldName, Direction.Ascending)];
-        break;
-    }
-
-    const newConfig = config.toBuilder().sort(newSort).build();
-    onConfigChange(newConfig).then(() => {
-      setLoadingState(false);
-    });
-  }
 
   _columnStyle = (fieldName: string) => {
     const { fields } = this.props;
@@ -222,31 +183,6 @@ class MessageTable extends React.Component<Props, State> {
   _fieldTypeFor = (fieldName: string, fields: Immutable.List<FieldTypeMapping>) => {
     return ((fields && fields.find(f => f.name === fieldName)) || { type: FieldType.Unknown }).type;
   };
-
-  _fieldSortDirection = (config: MessagesWidgetConfig, fieldName: string) => {
-    const currentSort = !isEmpty(config.sort) ? config.sort[0] : null;
-    if (currentSort && currentSort.field === fieldName) {
-      return currentSort.direction;
-    }
-    return undefined;
-  }
-
-  _fieldSortDescription = (fielName: string, fieldSortDirection?: Direction) => {
-    let newDirection;
-    if (fieldSortDirection && fieldSortDirection.direction === Direction.Ascending) {
-      newDirection = Direction.Descending.direction;
-    } else {
-      newDirection = Direction.Ascending.direction;
-    }
-    return `Sort ${fielName} ${newDirection}`;
-  }
-
-  _fieldSortIcon = (fieldSortDirection?: string) => {
-    if (fieldSortDirection && fieldSortDirection === Direction.Descending) {
-      return 'sort-amount-desc';
-    }
-    return 'sort-amount-asc';
-  }
 
   _getFormattedMessages = (): Array<Object> => {
     const { messages } = this.props;
@@ -279,7 +215,7 @@ class MessageTable extends React.Component<Props, State> {
 
   render() {
     const { expandedMessages } = this.state;
-    const { fields, activeQueryId, config, editing } = this.props;
+    const { fields, activeQueryId, config, editing, onConfigChange, setLoadingState } = this.props;
     const formattedMessages = this._getFormattedMessages();
     const selectedFields = this._getSelectedFields();
     return (
@@ -288,8 +224,6 @@ class MessageTable extends React.Component<Props, State> {
           <TableHead>
             <tr>
               {selectedFields.toSeq().map((selectedFieldName) => {
-                const fieldSortDirection = this._fieldSortDirection(config, selectedFieldName);
-                const fieldSortDescription = this._fieldSortDescription(selectedFieldName, fieldSortDirection);
                 return (
                   <th key={selectedFieldName}
                       style={this._columnStyle(selectedFieldName)}>
@@ -297,12 +231,10 @@ class MessageTable extends React.Component<Props, State> {
                            name={selectedFieldName}
                            queryId={activeQueryId} />
                     {editing && (
-                      <SortIcon fieldSortActive={!!fieldSortDirection}
-                                title={fieldSortDescription}
-                                aria-label={fieldSortDescription}
-                                onClick={() => this._changeSort(config, selectedFieldName, fieldSortDirection)}>
-                        <Icon name={this._fieldSortIcon(fieldSortDirection)} />
-                      </SortIcon>
+                      <MessagesSortIcon config={config}
+                                        onConfigChange={onConfigChange}
+                                        setLoadingState={setLoadingState}
+                                        fieldName={selectedFieldName} />
                     )}
                   </th>
                 );
