+++ graylog2-web-interface/src/views/components/widgets/MessagesSortIcon.jsx
@@ -1,0 +1,97 @@
+// @flow strict
+import React from 'react';
+import styled, { css, type StyledComponent } from 'styled-components';
+import { isEmpty } from 'lodash';
+
+import MessagesWidgetConfig from 'views/logic/widgets/MessagesWidgetConfig';
+import Direction from 'views/logic/aggregationbuilder/Direction';
+import SortConfig from 'views/logic/aggregationbuilder/SortConfig';
+
+import { Icon } from 'components/common';
+
+type Props = {
+    config: MessagesWidgetConfig,
+    fieldName: string,
+    onConfigChange: (MessagesWidgetConfig) => Promise<void>,
+    setLoadingState: (loading: boolean) => void,
+}
+
+const SortIcon: StyledComponent<{fieldSortActive: boolean}, {}, HTMLButtonElement> = styled.button(({ fieldSortActive }) => {
+  const color = fieldSortActive ? '#333' : '#bdbdbd';
+
+  return css`
+      border: 0;
+      background: transparent;
+      color: ${color};
+  
+      padding: 5px;
+  
+      cursor: pointer;
+    `;
+});
+
+const _changeSort = (config: MessagesWidgetConfig, fieldName: string, fieldSortDirection: ?Direction, onConfigChange: (MessagesWidgetConfig) => Promise<void>, setLoadingState: (loading: boolean) => void) => {
+  let newSort;
+
+  setLoadingState(true);
+
+  switch (fieldSortDirection) {
+    case Direction.Ascending:
+      newSort = config.sort.map(sort => sort.toBuilder().direction(Direction.Descending).build());
+      break;
+    case Direction.Descending:
+      newSort = config.sort.map(sort => sort.toBuilder().direction(Direction.Ascending).build());
+      break;
+    default:
+      newSort = [new SortConfig('pivot', fieldName, Direction.Ascending)];
+      break;
+  }
+
+  const newConfig = config.toBuilder().sort(newSort).build();
+  onConfigChange(newConfig).then(() => {
+    setLoadingState(false);
+  });
+};
+
+
+const _fieldSortDirection = (config: MessagesWidgetConfig, fieldName: string) => {
+  const currentSort = !isEmpty(config.sort) ? config.sort[0] : null;
+  if (currentSort && currentSort.field === fieldName) {
+    return currentSort.direction;
+  }
+  return undefined;
+};
+
+const _fieldSortDescription = (fielName: string, fieldSortDirection?: Direction) => {
+  let newDirection;
+  if (fieldSortDirection && fieldSortDirection.direction === Direction.Ascending) {
+    newDirection = Direction.Descending.direction;
+  } else {
+    newDirection = Direction.Ascending.direction;
+  }
+  return `Sort ${fielName} ${newDirection}`;
+};
+
+const _sortIcon = (fieldSortDirection?: string) => {
+  if (fieldSortDirection && fieldSortDirection === Direction.Descending) {
+    return 'sort-amount-desc';
+  }
+  return 'sort-amount-asc';
+};
+
+
+const MessagesSortIcon = ({ fieldName, config, onConfigChange, setLoadingState }: Props) => {
+  const fieldSortDirection = _fieldSortDirection(config, fieldName);
+  const fieldSortDescription = _fieldSortDescription(fieldName, fieldSortDirection);
+  return (
+    <SortIcon fieldSortActive={!!fieldSortDirection}
+              title={fieldSortDescription}
+              aria-label={fieldSortDescription}
+              onClick={() => _changeSort(config, fieldName, fieldSortDirection, onConfigChange, setLoadingState)}>
+      <Icon name={_sortIcon(fieldSortDirection)} />
+    </SortIcon>
+  );
+};
+
+
+export default MessagesSortIcon;
